I separated the functions in another file and made an import
